+++ src/variant.order.spec.ts
@@ -1,0 +1,34 @@
+import {CompareResult, sequence} from './sequence';
+import {Animal, cerberus} from './__test__/animal';
+
+const rank = sequence(Animal, [
+    'dog',
+    'cat',
+    Animal.snake,
+])
+
+const perseus = rank.new.cat({name: 'Perseus', furnitureDamaged: 0});
+
+test('order new obj', () => {
+    expect(perseus.name).toBe('Perseus');
+    expect(perseus.furnitureDamaged).toBe(0);
+})
+
+test('order compare', () => {
+    expect(rank.compare(perseus, cerberus)).toBe(CompareResult.Greater);
+})
+
+test('order compare', () => {
+    expect(rank.compare(Animal.cat, cerberus)).toBe(CompareResult.Greater);
+})
+
+test('order index', () => {
+    expect(rank.index(cerberus)).toBe(0);
+    expect(rank.index('cat')).toBe(1);
+    expect(rank.index(Animal.snake)).toBe(2);
+})
+
+test('get', () => {
+    expect(rank.get(0).type).toBe('dog');
+    expect(rank.types[0]).toBe('dog');
+})